% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_womens_data_2020}
\alias{import_womens_data_2020}
\title{Imports data for women's marathon from the 2020 Tokyo Olympics!}
\usage{
import_womens_data_2020()
}
\value{
a tibble with the ranked list of women's Olympic marathon runners from the 2020 Tokyo Olympics and associated times
}
\description{
Imports data for women's marathon from the 2020 Tokyo Olympics!
}
\examples{
import_womens_data_2020()
}
