% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mens_2008_data}
\alias{mens_2008_data}
\title{\code{mens_2008_data} contains data from the 2008 Beijing Olympics Men's Marathon}
\format{
\code{mens_2008_data} is an object of class tbl_df (inherits from tbl, data.frame) with 98 rows and 9 columns
}
\source{
\code{mens_2008_data} comes from the \href{https://www.worldathletics.org/results/olympic-games/2008/the-xxix-olympic-games-6977748/men/marathon/final/result}{World Athletics website's 2008 men's marathon results page}
}
\usage{
mens_2008_data
}
\description{
\code{mens_2008_data} contains data from the 2008 Beijing Olympics Men's Marathon
}
\examples{
mens_2008_data
}
\keyword{datasets}
