% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times_over_time.R
\name{times_over_time}
\alias{times_over_time}
\title{Shows finishing time data over time.}
\usage{
times_over_time()
}
\value{
ggplot object
}
\description{
Creates a ggplot visualization of all complete finishing times in \code{olympic_marathon}
}
\examples{
times_over_time()

}
