% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_country_viz.R
\name{two_country_viz}
\alias{two_country_viz}
\title{Visualization of finishing times from two countries}
\usage{
two_country_viz(nationality1, nationality2)
}
\arguments{
\item{nationality1}{3 character indicator for first nationality}

\item{nationality2}{3 character indicator for second nationality}
}
\value{
ggplot object
}
\description{
Creates a ggplot of the finishing times in \code{olympic_marathon} of all competitors from two input countries (excluding DNF)
}
\examples{
two_country_viz("ITA", "NOR")

}
