% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{womens_2004_data}
\alias{womens_2004_data}
\title{\code{womens_2004_data} contains data from the 2004 Athens Olympics Women's Marathon}
\format{
\code{womens_2004_data} is an object of class tbl_df (inherits from tbl, data.frame) with 82 rows and 9 columns
}
\source{
\code{womens_2004_data} comes from the \href{https://www.worldathletics.org/results/olympic-games/2994/the-xxvii-olympic-games-6913163/women/marathon/final/result}{World Athletics website's 2004 women's marathon results page}
}
\usage{
womens_2004_data
}
\description{
\code{womens_2004_data} contains data from the 2004 Athens Olympics Women's Marathon
}
\examples{
womens_2004_data
}
\keyword{datasets}
