% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_womens_data_2020}
\alias{import_womens_data_2020}
\title{Imports data for women's marathon from 2020!}
\usage{
import_womens_data_2020()
}
\description{
Imports data for women's marathon from 2020!
}
