% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mens_2020_data}
\alias{mens_2020_data}
\title{This file contains data from 2020 Tokyo Olympics Men's Marathon!}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 106 rows and 9 columns.
}
\usage{
mens_2020_data
}
\description{
This file contains data from 2020 Tokyo Olympics Men's Marathon!
}
\keyword{datasets}
