% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{womens_2020_data}
\alias{womens_2020_data}
\title{This file contains data from 2020 Tokyo Olympics Women's Marathon!}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 88 rows and 9 columns.
}
\usage{
womens_2020_data
}
\description{
This file contains data from 2020 Tokyo Olympics Women's Marathon!
}
\keyword{datasets}
