% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{womens_2020_data}
\alias{womens_2020_data}
\alias{mens_2020_data}
\alias{womens_2016_data}
\alias{mens_2016_data}
\alias{womens_2012_data}
\alias{mens_2012_data}
\title{Olympic Marathon Data by Gender and Year}
\format{
\code{womens_2020_data} is an object of class tbl_df (inherits from tbl, data.frame) with 88 rows and 9 columns
\describe{
\item{rank}{Finishing place based on race time}
\item{name}{Competitor's Name}
\item{nationality}{3 letter abbreviation of the country that the competitor is representing}
\item{result}{The competitor's finishing time (hr:min:sec) (DNF if did not finish, DNS if did not start, DQ if disqualified)}
\item{gender}{Gender of the competitor (F = female, M = male)}
\item{event}{The Olympic event that the entry corresponds to}
\item{location}{The location of the Olympic Games}
\item{year}{The year of the Olympic Games}
\item{medal}{What medal the competitor earned, if any (G = Gold, S = Silver, B = Bronze)}
}

\code{mens_2020_data} is an object of class tbl_df (inherits from tbl, data.frame) with 106 rows and 9 columns

\code{womens_2016_data} is an object of class tbl_df (inherits from tbl, data.frame) with 157 rows and 9 columns

\code{mens_2016_data} is an object of class tbl_df (inherits from tbl, data.frame) with 155 rows and 9 columns

\code{womens_2012_data} is an object of class tbl_df (inherits from tbl, data.frame) with 115 rows and 9 columns

\code{mens_2012_data} is an object of class tbl_df (inherits from tbl, data.frame) with 105 rows and 9 columns
}
\source{
\code{womens_2020_data} comes from the \href{https://www.worldathletics.org/results/olympic-games/2021/the-xxxii-olympic-games-athletics-7132391/women/marathon/final/result}{World Athletics website's 2020 women's marathon results page}

\code{mens_2020_data} comes from the \href{https://www.worldathletics.org/results/olympic-games/2021/the-xxxii-olympic-games-athletics-7132391/men/marathon/final/result}{World Athletics website's 2020 men's marathon results page}

\code{womens_2016_data} comes from the \href{https://www.worldathletics.org/results/olympic-games/2016/the-xxxi-olympic-games-7093747/women/marathon/final/result}{World Athletics website's 2016 women's marathon results page}

\code{mens_2016_data} comes from the \href{https://www.worldathletics.org/results/olympic-games/2016/the-xxxi-olympic-games-7093747/men/marathon/final/result}{World Athletics website's 2016 men's marathon results page}

\code{womens_2012_data} comes from the \href{https://www.worldathletics.org/results/olympic-games/2012/the-xxx-olympic-games-6999193/women/marathon/final/result}{World Athletics website's 2012 women's marathon results page}

\code{mens_2012_data} comes from the \href{https://www.worldathletics.org/results/olympic-games/2012/the-xxx-olympic-games-6999193/men/marathon/final/result}{World Athletics website's 2012 men's marathon results page}
}
\usage{
womens_2020_data

mens_2020_data

womens_2016_data

mens_2016_data

womens_2012_data

mens_2012_data
}
\description{
\code{womens_2020_data} contains data from the 2020 Tokyo Olympics Women's Marathon

\code{mens_2020_data} contains data from the 2020 Tokyo Olympics Men's Marathon

\code{womens_2016_data} contains data from the 2016 Rio de Janeiro Olympics Women's Marathon

\code{mens_2016_data} contains data from the 2016 Rio de Janeiro Olympics Men's Marathon

\code{womens_2012_data} contains data from the 2016 Rio de Janeiro Olympics Men's Marathon

\code{mens_2012_data} contains data from the 2016 Rio de Janeiro Olympics Men's Marathon
}
\examples{
womens_2020_data
mens_2020_data
womens_2016_data
mens_2016_data
womens_2012_data
mens_2012_data
}
\keyword{datasets}
